requirements.txt
pandas
requests
Pillow
PyYAML
reportlab
pytest
types-requests
types-PyYAML
